+++ docs/conf.py
@@ -11,10 +11,10 @@ sys.path.insert(0, os.path.abspath(".."))
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = "Royal Flush"
+project = "Royal FLush"
 copyright = "2024, Francisco Enguix"
 author = "Francisco Enguix"
-release = "0.1.0"
+release = "0.3.4"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration