+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+- Add PMACoFL Max algorithm.
+- Add parsing to Diritchlet alpha input.
+
 ## [0.3.3] - 2025-01-27
 
 ### Changed